Moves get_cal_obsids to load_book
